backend/requirements.txt
alembic==1.8.1
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.88.0
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
